feat(viewer): gunicorn-based viewer (#18)
For whatever, reason, using `mod_wsgi` is the cause of #13 so here's a
version designed to run in `gunicorn` inside a virtualenv. (`uwsgi`
would presumably also work.)

Installation instructions have been updated in the README, but there's
also a `ch_ansible` play (or will be once I get it working...)

For ease of use, I've created two bash scripts:
* `deploy_to_venv` can be used to install your local check-out into the
production venv, meaning you don't have to run the full ch_ansible play
to deploy, if you have a local check-out on bao. The gunicorn virtualenv
must already be set up (per the README and/or ch_ansible play).
* `test_server` can be used to set-up the environment for, and then run
a test-version of the server that will be accessible at the `daily-test`
URL. Run this script in a `screen` or something for maximal ease-of-use.
This script must be run from from within a gunicorn virtualenv, either
the production version, or one you've rolled yourself.

## Itemised changes

* A `requirements.txt` file to list requirements (used by `ch_ansible`)
* Deleted the `mod_wsgi` boilerplate `web/viewer.wsgi`
* No longer use `HTTP_HOST` for URL reconstruction, which apache sets to
the proxying hostname.
* HTML5 fixes to the templates
* The viewer now sets both the `Content-Length` header and the `charset`
agrument to the `Content-Type` header for better HTTP compliance, though
I don't think either was the cause of the trouble we were seeing.
* Session timeout is now 7 days per @jrs65 's request. (See the
`SESSION_EXPIRY_DAYS` constant).
* The `rendered` directory has been removed from the git repository. It
needs to be created as part of the installation procedure.
* Fixed a bug in some code in `daily.js` which should not normally be
reachable. (Unless the user is injecting their own JS into the viewer,
which they're technically allowed to do.)
* Moved `fetch_from_cedar.sh` to `ch_ansible` so it can be properly
templated.

Closes #13