feat(viewer): Revision selection and back-end rewrite (#25)
This is a fairly extensive re-write with the following goals:
* make pipeline revision both dynamic and selectable
* remove the front-loading of _all_ a user's opinions on first log-in

Additionally, there's been a bit of tidying up

## Removing front-loading

Closes #19 

As first written, the viewer used to find all opinions for a user when
they first logged in. This data was shoved into the rendered HTML
template (in a very non-JSON-y way, for unknown reasons). The only
benefit of doing this was that the client never had to contact the
server again (except if the user wanted to update their opinion).

Fine in theory, if there are only a few opinions for a user (which was
true when the viewer was first written), but causes more-and-more
slowdown as opinions accumulate over time.

In place of this, the client now side-loads opinions for a CSD (for all
available revisions) by sending a `fetch` request to the server, which
returns the data in JSON (see `fetch_csd` in the server, though most of
the work is done in `csd_data`; and also `set_csd` in the client).

The downside here is there is a delay when switching CSDs in the viewer
as the viewer has to wait for the server to return the data, but I don't
find this delay noticable in practice.

Specific server-side changes:
* NB: the "fetch" query parameter was previously used when the client
asked for a rendered notebook page. That key has been renamed to
"render". The "fetch" parameter is now used for a CSD data side-load
request
* The `fetch_csd` function is used to respond to these side-loading
requests. Most of the actual data generation is implemented in
`csd_data` (which evolved from the `csd_vars` function and is also used
after an opinion update to generate the data returned to the client).

Specific client-side changes:
* The `csd_list` global has been deleted (formerly a list of all
available CSDs)
* The `opinions` global now holds the user's opinion for the current CSD
only (indexed by revision), rather than all CSDs.
* The `vet_csd` function, which was used to validate numbers typed into
the CSD box, has been deleted (the functionality is now handled by the
server during the side-loading)

## Pipeline revision selection

Closes #20

The fixed global `_REVISION` has been removed. In its place, the viewer
now will dynamically figure out which revisions to load/display:

* On first load (i.e. when a worker first serves a page and loads its
globals, the server walks through `DataRevision` and finds all revisions
for which at least one rendered notebook exists. This ends up as the
`REVISIONS` global, and will remain fixed until server restart.
* Due to the side-loading, whenever data from a new CSD is requested,
the server will find all rendered notebooks for that CSD and provide the
data to the client.
* In the client there are now buttons to switch between revisions. The
list of buttons is fixed by the contents of `REVISIONS` (i.e. it doesn't
change unless a server restart happens) but, if the currently-viewed CSD
has no rendered notebook for a particular revision, the corresponding
revision button is disabled to indicate that.
* Otherwise, revision buttons are coloured to indicate the user's
submitted opinion for that day (if any).
* Because a CSD-fetch side-load returns all revisions, switching between
revisions for a CSD does not cause another server hit.

Specific changes:
* revision number (`rev` parameter) must be submitted with all requests
now, and is provided in all responses. When not given a valid revision
number, the server will default to the lowest available revision.
Instances where a client's query data was stripped of "unimportant" data
has now been unified into the `strip_query` function.
* The `update_ui` client function has been broken up into two parts:
`update_data` which updates the globals based on a server's response,
and `draw_ui` which performs the document update based on the state of
the globals. `update_data` is called by side-loader callbacks, and in
turn will call `draw_ui`, but the `draw_ui` function is also called by
the revision selection buttons (via their `set_rev` function), which
doesn't involve a side-load.

## Other changes

I've removed the `script_name` and `base_path` vars which were an
unnecessary attempt to tell the client where to send requests: the
client doesn't need this; it already knows the URL it's using. More
evidence that these are not needed: they were broken when we
transitioned the viewer to gunicorn (both ended up as `""`) but the
client didn't notice or care).

This PR breaks up the `update_opinion` function in the server, which
isn't technically necessary, and only done in anticipation of the
category handling stuff.