feat(viewer): remember chosen revision (#34)
Well, this was easier than I had anticipated.

This is a client-only change:

When a user clicks on a revision button, the revision they clicked on is
now remembered (as `prefrev`) and is used always when making `fetch`
side-load requests. The net effect is, in cases when the server changes
the revision (because there was no render for the starting revision),
the client will use the changed revision (stored in `rev`) for rendering
pages until the starting revision is again available for a CSD, at which
point it will switch back.

Also cleaned up the console log a bit.